move legacy ButtonText
